MDEV-14975 mariabackup is starting unprivileges user.
Ported basic privilege checks from percona xtrabackup.
The backup could be a long process, and it would be pity to recognize
towards the end of backup that required privilege is not held.

Added an option --check-privileges(default ON).
When this option is specified, backup would check for RELOAD privilege
required by FLUSH TABLES WITH READ LOCK, and exit if user does not have it.

Additionally, SUPER privilege is checked in cases backup can issue
KILL QUERY, and also REPLICATION CLENT for SHOW MASTER/SLAVE STATUS.
Warnings will be given if backup does not have the privilege.